Conflicts:
	src/main/resources/egovframework/sqlmap/ezEKP/init_data_mysql.sql
	src/main/resources/egovframework/sqlmap/ezEKP/init_data_oracle.sql

+++ src/main/webapp/WEB-INF/jsp/ezEmail/mailLeft.jsp
@@ -1,6 +1,4 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<%@ page import="egovframework.let.utl.fcc.service.CommonUtil" %>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@ taglib uri="http://www.springframework.org/tags" prefix="spring" %>
 <!DOCTYPE html>
@@ -8,19 +6,19 @@
 	<head>
 	    <title>left_myoffice</title>
 	    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-		<script type="text/javascript" src="<%=CommonUtil.addVer(application, request, "<spring:message code='ezEmail.e1' />")%>"></script>
-	    <script type="text/javascript" src="<%=CommonUtil.addVer(application, "/js/mouseeffect.js")%>"></script>
-	    <script type="text/javascript" src="<%=CommonUtil.addVer(application, "/js/jquery/jquery-1.11.3.min.js")%>"></script>
-	    <script type="text/javascript" src="<%=CommonUtil.addVer(application, "/js/ezEmail/js_cross/email_tree.js")%>"></script>
-	    <script type="text/javascript" src="<%=CommonUtil.addVer(application, "/js/ezEmail/Controls_cross/treeview.htc.js")%>"></script>
-	    <script type="text/javascript" src="<%=CommonUtil.addVer(application, "/js/ezEmail/js_cross/string_component_utf8.js")%>"></script>
-	    <script type="text/javascript" src="<%=CommonUtil.addVer(application, "/js/ezEmail/js_cross/encode_component.js")%>"></script>
-	    <script type="text/javascript" src="<%=CommonUtil.addVer(application, "/js/XmlHttpRequest.js")%>"></script>
-	    <script type="text/javascript" src="<%=CommonUtil.addVer(application, "/js/ezAddress/address_tree_Cross.js")%>"></script>
+		<script type="text/javascript" src="${util.addVer('ezEmail.e1', 'msg')}"></script>
+	    <script type="text/javascript" src="${util.addVer('/js/mouseeffect.js')}"></script>
+	    <script type="text/javascript" src="${util.addVer('/js/jquery/jquery-1.11.3.min.js')}"></script>
+	    <script type="text/javascript" src="${util.addVer('/js/ezEmail/js_cross/email_tree.js')}"></script>
+	    <script type="text/javascript" src="${util.addVer('/js/ezEmail/Controls_cross/treeview.htc.js')}"></script>
+	    <script type="text/javascript" src="${util.addVer('/js/ezEmail/js_cross/string_component_utf8.js')}"></script>
+	    <script type="text/javascript" src="${util.addVer('/js/ezEmail/js_cross/encode_component.js')}"></script>
+	    <script type="text/javascript" src="${util.addVer('/js/XmlHttpRequest.js')}"></script>
+	    <script type="text/javascript" src="${util.addVer('/js/ezAddress/address_tree_Cross.js')}"></script>
 	    <!-- 재은 수정 -->
-	    <script type="text/javascript" src="<%=CommonUtil.addVer(application, "/js/ezEmail/js_cross/NewMailList.js")%>"></script>
-	    <link rel="stylesheet" href="<%=CommonUtil.addVer(application, request, "<spring:message code='ezEmail.c1' />")%>" type="text/css">
-	    <link rel="stylesheet" href="<%=CommonUtil.addVer(application, request, "<spring:message code='main.lhm02' />")%>" type="text/css">
+	    <script type="text/javascript" src="${util.addVer('/js/ezEmail/js_cross/NewMailList.js')}"></script>
+	    <link rel="stylesheet" href="${util.addVer('ezEmail.c1', 'msg')}" type="text/css">
+	    <link rel="stylesheet" href="${util.addVer('main.lhm02', 'msg')}" type="text/css">
 	    <script type="text/javascript">
 	        var pUse_Editor = "${useEditor}";
 	        var subCode = "${subCode}";
@@ -862,13 +860,17 @@
 			      	//지운편지함의 메일 영구삭제
 			        if (folderPath == trashBoxURL) {
 			            if (confirm("<spring:message code='ezEmail.t470' />")) {
-			                delete_mail(folderPath, true, "");
+				            if (confirm("<spring:message code='ezEmail.ksa03' />")) {
+				                delete_mail(folderPath, true, "");
+				            }
 			            }
 			        }
-			      	//편지함의 메일 지운편지함으로 이동
+			      	//편지함의 메일 지운편지함으로 이동  
 			        else {
 			            if (confirm("<spring:message code='ezEmail.t475' />")) {
-			                delete_mail(folderPath, false, trashBoxURL);
+			            	if (confirm("<spring:message code='ezEmail.ksa04' />")) {
+				                delete_mail(folderPath, false, trashBoxURL);	
+			            	}
 			            }
 			        }
